25.05.2026 Perenos

+++ code/datums/components/_component.dm
@@ -49,8 +49,12 @@
 /datum/component/New(list/raw_args)
 	parent = raw_args[1]
 	var/list/arguments = raw_args.Copy(2)
-	if(Initialize(arglist(arguments)) == COMPONENT_INCOMPATIBLE)
+	var/init_result = Initialize(arglist(arguments))
+	if(init_result == COMPONENT_INCOMPATIBLE)
 		stack_trace("Incompatible [type] assigned to a [parent.type]! args: [json_encode(arguments)]")
+		qdel(src, TRUE, TRUE)
+		return
+	if(init_result == COMPONENT_INCOMPATIBLE_SILENT)
 		qdel(src, TRUE, TRUE)
 		return
